chore: improved module structure
- use separte modules for individual concepts
- avoid putting code to __init__.py
