+++ wlc/const.py
@@ -1,0 +1,15 @@
+# Copyright © Michal Čihař <michal@weblate.org>
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+"""Package constants."""
+
+VERSION = "1.17.2"
+
+URL = "https://weblate.org/"
+DEVEL_URL = "https://github.com/WeblateOrg/wlc"
+API_URL = "http://127.0.0.1:8000/api/"
+USER_AGENT = f"wlc/{VERSION}"
+LOCALHOST_ADDRESSES = {"127.0.0.1", "localhost", "::1", "[::1]"}
+
+TIMESTAMPS = {"last_change"}
